# dogfightapi/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Dog Fight App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
schemes:  
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /hello:
    # binds a127 app logic to a route
    x-swagger-router-controller: hello_world
    get:
      description: Health check
      # used as the method name of the controller
      operationId: hello      
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            type: string
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /dogs:
    x-swagger-router-controller: dog
    get:
      description: List all dogs
      operationId: getDogs
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Dogs"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /dog:
    x-swagger-router-controller: dog
    post:
      description: Add a dog to the fight
      operationId: addDog
      parameters:
        - name: dog
          in: body
          description: dog information
          required: true
          schema:
            $ref: "#/definitions/Dog"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /dog/{name}:
    x-swagger-router-controller: dog
    parameters:
      - name: name
        in: path
        description: Name of dog
        required: true
        type: string
    get:
      description: Returns dog details
      operationId: getDog
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Dog"      
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: Remove a dog
      operationId: removeDog
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /dog/{name}/attacks:
    x-swagger-router-controller: dog
    parameters:
      - name: name
        in: path
        description: The name of the dog to remove
        required: true
        type: string
    get:
      description: Gets dogs's fight moves
      operationId: getAttacks
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Attacks"      
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Adds an attack for the dog
      operationId: addAttack
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"      
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /dog/{name}/distractions:
    x-swagger-router-controller: dog
    parameters:
      - name: name
        in: path
        description: The name of the dog to remove
        required: true
        type: string
    get:
      description: Gets dogs's distractions
      operationId: getDistractions
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Distractions"      
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Adds a distraction for the dog
      operationId: addDistraction
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"      
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /dog/{name}/actions:
    x-swagger-router-controller: dog
    parameters:
      - name: name
        in: path
        description: The name of the dog to remove
        required: true
        type: string
    get:
      description: Gets dogs's actions
      operationId: getActions
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Actions"      
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Adds an action for the dog
      operationId: addAction
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"      
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /fight:
    x-swagger-router-controller: fight
    post:
      description: Start a dog fight
      operationId: start
      parameters:
        - name: dogs
          in: body
          description: Add 2 dogs to a fight
          required: true
          schema:
            $ref: "#/definitions/DogFight"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /listen:
    x-swagger-router-controller: listen    
    post:
      consumes: 
        - application/x-www-form-urlencoded
      description: Listens for commands
      operationId: listen
      parameters:
        - $ref: "#/parameters/token"
        - $ref: "#/parameters/team_id"
        - $ref: "#/parameters/team_domain"
        - $ref: "#/parameters/channel_id"
        - $ref: "#/parameters/channel_name"
        - $ref: "#/parameters/user_id"
        - $ref: "#/parameters/user_name"
        - $ref: "#/parameters/text"
        - $ref: "#/parameters/trigger_word"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SlackResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /swagger:
    x-swagger-pipe: swagger_raw
parameters:
  token:
    name: token
    in: formData
    description: token
    required: true
    type: string
  team_id:
    name: team_id
    in: formData
    description: team_id
    type: string
  team_domain:
    name: team_domain
    in: formData
    description: team_domain
    type: string
  channel_id:
    name: channel_id
    in: formData
    description: channel_id
    type: string
  channel_name:
    name: channel_name
    in: formData
    description: channel_name
    type: string
  user_id:
    name: user_id
    in: formData
    description: user_id
    type: string
  user_name:
    name: user_name
    in: formData
    description: user_name
    type: string
  text:
    name: text
    in: formData
    description: text
    type: string
  trigger_word:
    name: trigger_word
    in: formData
    description: trigger_word
    type: string  
# complex objects have schema definitions
definitions:
  SlackResponse:
    type: object
    properties:      
      text:
        type: string
        description: Message to send to Slack
  Actions:
    type: array
    items:
      type: object
      properties:
        action:
          type: string
        value:
          type: integer
          format: int32
  Distractions:
    type: array
    items: 
      type: object
      properties:
        distraction:
          type: string
        value:
          type: integer
          format: int32
  Attacks:
    type: array
    items:
      type: object
      properties:
        verb:
          type: string
        noun:
          type: string
        value:
          type: integer
          format: int32
  DogFight:
    type: array
    items:
      type: string
  Dogs:
    type: array
    items:
      type: string
  GeneralResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      description:
        type: string
        description: a short comment 
    required:
      - success
      - description
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  Dog:
    required:
      - name
      - weight
    properties:
      name:
        type: string
      gender:
        type: string
      weight:
        type: integer
        format: int32
      attacks:
        $ref: "#/definitions/Attacks"
      distractions:
        $ref: "#/definitions/Distractions"
      actions:
        $ref: "#/definitions/Actions"
      
            
            
      
      